scroll [nfc]: Give SingleChildScrollViewWithScrollbar docs and a new home
